requirements.txt
discord.py==1.7.1
aiohttp==3.7.4.post0
SQLAlchemy==1.4.6
asyncpg==0.22.0